+++ simbot-cores/simbot-core-spring-boot-starter-common/src/main/java/module-info.java
@@ -5,6 +5,7 @@ module simbot.spring.common {
     requires transitive simbot.core;
     requires static java.annotation;
     requires kotlinx.coroutines.core;
+    requires kotlinx.serialization.core;
 
     exports love.forte.simbot.spring.common.application;
     exports love.forte.simbot.spring.common;