+++ simbot-cores/simbot-core/src/jvmMain/java/module-info.java
@@ -4,6 +4,7 @@ module simbot.core {
     requires static simbot.common.annotations;
     requires transitive simbot.common.collection;
     requires kotlinx.coroutines.core;
+    requires kotlinx.serialization.core;
 
     exports love.forte.simbot.core.application;
     exports love.forte.simbot.core.event;